score

+++ db/migrate/20190805102929_create_post_times.rb
@@ -1,7 +1,7 @@
 class CreatePostTimes < ActiveRecord::Migration[5.2]
   def change
     create_table :post_times do |t|
-      t.integer :study_time
+      t.float :study_time
       t.integer :user_id
 
       t.timestamps
